requirements.txt
flask
openai-whisper
transformers
datasets
torch
soundfile
numpy
jiwer
